requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
azure-storage-blob==12.19.0
schedule==1.2.0
